+++ wn/_config.py
@@ -14,7 +14,7 @@ from wn._util import is_url, resources, short_hash
 
 # The directory where downloaded and added data will be stored.
 DEFAULT_DATA_DIRECTORY = Path.home() / '.wn_data'
-DEFAULT_DATABASE_FILENAME = 'wn.db'
+DATABASE_FILENAME = 'wn.db'
 
 
 class WNConfig:
@@ -22,7 +22,7 @@ class WNConfig:
     def __init__(self):
         self._data_directory = DEFAULT_DATA_DIRECTORY
         self._projects = {}
-        self.database_filename = DEFAULT_DATABASE_FILENAME
+        self._dbpath = self._data_directory / DATABASE_FILENAME
 
     @property
     def data_directory(self) -> Path:
@@ -37,11 +37,12 @@ class WNConfig:
         if dir.exists() and not dir.is_dir():
             raise Error(f'path exists and is not a directory: {dir}')
         self._data_directory = dir
+        self._dbpath = dir / DATABASE_FILENAME
 
     @property
     def database_path(self):
         """The path to the database file."""
-        return self.data_directory / self.database_filename
+        return self._dbpath
 
     @property
     def downloads_directory(self):